Dataset: Fix Negative Values
Negative integers lowered the confidence of the predictions. Now the lowest possible negative integer is -5.
Note: Input image has been changed to a more realistic input image of the cassiopeia constellation. This constellation is not supported with the program, hence running the prediction will give incorrect result with this input.